Move NetworkMetrics definition
